server/requirements.txt
django
djangorestframework
pillow
pyjwt
django-cors-headers
